+++ routes/routes.php
@@ -1,5 +1,8 @@
 <?php
-require_once('lib/db.php');
+header('Access-Control-Allow-Origin: *');
+header('Content-Type: application/json');
+header('Access-Control-Allow-Methods: GET');
+header('Access-Control-Allow-Headers: Access-Control-Allow-Headers,Content-Type,Access-Control-Allow-Methods,Authorization,X-Requested-With');
 
 
 $routes = explode("/",$_SERVER["REQUEST_URI"]);
